Add phase-out within a range and refactor
