+++ openfisca_us/parameters/contrib/ubi_center/basic_income/agi_limit/amount.yaml
@@ -1,14 +1,14 @@
 description: Tax units are eligible for basic income payments unless their adjusted gross income exceeds this threshold.
 SINGLE:
-  2010-01-01: 0
+  0000-01-01: 0
 JOINT:
-  2010-01-01: 0
+  0000-01-01: 0
 HEAD_OF_HOUSEHOLD:
-  2010-01-01: 0
+  0000-01-01: 0
 SEPARATE:
-  2010-01-01: 0
+  0000-01-01: 0
 WIDOW:
-  2010-01-01: 0
+  0000-01-01: 0
 metadata:
   unit: currency-USD
   period: year
