+++ openfisca_us/parameters/contrib/ubi_center/basic_income/agi_limit/in_effect.yaml
@@ -1,7 +1,7 @@
 description: Basic income is limited to tax units below an AGI level if this switch is activated.
 values:
-  2010-01-01: false
+  0000-01-01: false
 metadata:
-  unit: switch
+  unit: bool
   name: bi_agi_limit_in_effect
   label: Basic income AGI limit in effect
