+++ openfisca_us/parameters/contrib/ubi_center/basic_income/phase_out/end.yaml
@@ -1,4 +1,4 @@
-description: The basic income phases out for adjusted gross income above this value.
+description: The basic income phases out for adjusted gross income between the threshold and this value.
 SINGLE:
   0000-01-01: 0
 JOINT:
@@ -12,6 +12,6 @@ WIDOW:
 metadata:
   unit: currency-USD
   period: year
-  name: bi_phase_out_threshold
-  label: Basic income phase-out threshold
+  name: bi_phase_out_end
+  label: Basic income phase-out end
   breakdown: filing_status
