+++ openfisca_us/parameters/contrib/ubi_center/basic_income/phase_out/rate.yaml
@@ -1,6 +1,6 @@
-description: The rate at which basic income is recollected from taxpayers over the AGI threshold.
+description: The basic income phases out at this rate for tax filers with adjusted gross income above the threshold.
 values:
-  2010-01-01: 0
+  0000-01-01: 0
 metadata:
   unit: /1
   name: bi_phase_out_rate
