+++ openfisca_us/tests/policy/baseline/contrib/ubi_center/basic_income.yaml
@@ -2,18 +2,7 @@
   period: 2022
   absolute_error_margin: 1
   input:
-    contrib.ubi_center.basic_income.phase_out.rate: 0.1
-    contrib.ubi_center.basic_income.phase_out.threshold.HEAD_OF_HOUSEHOLD: 50_000
-    contrib.ubi_center.basic_income.amount_by_age[3].amount: 10_000
-    contrib.ubi_center.basic_income.amount_by_age[0].amount: 5_000
-    people:
-      adult:
-        employment_income: 100_000
-        age: 26
-      child:
-        age: 4
-    tax_units:
-      tax_unit:
-        members: [adult, child]
+    basic_income_before_phase_out: 15_000
+    basic_income_phase_out: 5_000
   output:
     basic_income: [6_667, 3_333]
