+++ openfisca_us/variables/contrib/ubi_center/basic_income/basic_income.py
@@ -8,16 +8,16 @@ class basic_income(Variable):
     unit = USD
     documentation = "Total basic income payments for this person. Phase-outs as an equal percentage to all tax unit members."
     definition_period = YEAR
+    defined_for = "basic_income_eligible"
 
     def formula(person, period, parameters):
-        basic_income = person("bi_before_phase_out", period)
+        basic_income = person("basic_income_before_phase_out", period)
         tax_unit = person.tax_unit
         tax_unit_basic_income = tax_unit.sum(basic_income)
-        tax_unit_phase_out = tax_unit("bi_phase_out", period)
+        tax_unit_phase_out = tax_unit("basic_income_phase_out", period)
         percent_reduction = where(
             tax_unit_basic_income > 0,
             tax_unit_phase_out / tax_unit_basic_income,
             0,
         )
-        eligible = tax_unit("basic_income_eligible", period)
-        return basic_income * (1 - percent_reduction) * eligible
+        return basic_income * (1 - percent_reduction)
