+++ openfisca_us/variables/contrib/ubi_center/basic_income/basic_income_before_phase_out.py
@@ -1,17 +1,17 @@
 from openfisca_us.model_api import *
 
 
-class bi_before_phase_out(Variable):
+class basic_income_before_phase_out(Variable):
     value_type = float
     entity = Person
     label = "Basic income before phase-outs"
     unit = USD
     documentation = (
-        "Total basic income payments for this person, before apply phase-outs."
+        "Total basic income payments for this person, before phasing out."
     )
     definition_period = YEAR
 
     def formula(person, period, parameters):
-        bi = parameters(period).contrib.ubi_center.basic_income
+        p = parameters(period).contrib.ubi_center.basic_income
         age = person("age", period)
-        return bi.amount_by_age.calc(age)
+        return p.amount_by_age.calc(age)
